requirements.txt
discord.py==2.3.2
pymongo==4.6.1
python-dotenv==1.0.0
easyocr==1.7.1
pillow==10.2.0
requests==2.31.0